[BUGFIX] tracing-limit: Drop state lock before calling inner layer
We encountered a deadlock where Thread A holds SystemState write lock
and tries to write a log message, going through tracing-limit dashmap
which tries to take the shard lock. This is while Thread B holds the
same shard lock, and in the inner layer a Valuable visit() method
attempts to lock SystemState read lock.

Change-Id: Iae640b928b8dfd9ff1b2caf0c84269849df52f69
Signed-off-by: Yonatan Goldschmidt <yonatan.goldschmidt@wiz.io>